Begin adding some SSE2 instructions.
I was going to try to get some software (LZSSE) working by implementing
only the necessary instructions.  While that would provide a nice way
to prioritize implementing instructions, I think it would be easier to
simply continue going in order; it may be possible to implement some
instructions by using other instructions from previous SSE versions.
I'm not familiar enough with the instruction sets to know which ones
could be used (which is part of the reason for this project), so I'll
just continue going in order, at least for now.