Improve symbol naming.
First, this allows you to define SIMDE_PREFIX to whatever you want,
the symbols will be generated as ${prefix}symbol.  This will allow
you to include SIMDe multiple times with different prefixes, which
should be useful for compilers which allow you to change targets
within a single compilation (i.e., non-MSVC).  Eventually I'd also
like to (ab)use this to create tests which compare native and emulated
versions using random inputs.

This also drops the old convention of using the preprocessor to define
"native" names to SIMDe names to provide an API compatible with
conventional intrinsics.  Instead, the default prefix is now "" and
everything (even wrappers) are generated as functions.

Due to the move from macros to functions, I found quite a few bugs
which this patch (hopefully) corrects.  I'd still feel more comfortable
with tests for wrapper functions since it's possible I've accidentally
mapped some functions incorrectly, but it's a lot of work for a
relatively small benefit.