Add native types to unions instead typedefing SIMDe types to them
This expands the code for wrapper functions a bit, since we have to
convert between native types and our own unions, but there shouldn't
be any impact on the generated machine code.

The main reason for this is that it provides a uniform API for
accessing data, both in code consuming the SIMDe API and other modules
(e.g., using MMX types in the SSE module).  Instead of having to use
the preprocessor to figure out if the data is using our unions or is
only a typedef to a native type, we can just access it through the
SIMDe union.