sse: work around clang problem with _mm_{extract,insert}_pi16
The functions require that the imm8 argument be a constant.  For GCC
it is sufficient to add the const keyword to the argument, and of
course the value passed to the wrapper must be a compile-time constant.
clang, OTOH, doesn't seem to allow a wrapper function at all... clang's
xmmintrin.h defines _mm_extract_pi16 and _mm_insert_pi16 as a macro to
get around this, but that's not an option for us since we need to be
able to generate different symbol names.

For now, we'll just always use the portable version on clang.
Suggestions for a more permanant fix welcome.