Remove ability to change symbol prefix.
This caused more problems than it was worth.  First, it made the
compiler output for warnings much harder to understand as you typically
had to wade through several layers of indirection.

More importantly, though perhaps less annoyingly, some functions simply
can't be implemented as functions in some compilers; `_mm_srli_si128`
is an example.

Changing the prefix was really only useful in certain situations.  I'm
particularly upset about loosing the ability to just run the native
functions against the portable versions from SIMDe with random data as
a test, but that is really only a problem for this project.

The big thing that projects *using* SIMDe loose is the ability to
include a header several times (with a different prefix) while changing
the target architecure each time, but not all compilers support this
*cough* MSVC *cough*.  The only real "portable" way to do this is to
compile a file multiple times with different options on the command
line, which SIMDE__SYMBOL didn't help with.