Replace float/double with new simde_float32/64 types
This should make it easier to port SIMDe to platform where float and
double aren't 32 and 64 bits, respectively.  I don't currently have
access to such a platform, so I'm not likely to be able to test this
any time soon, but I'd rather get the API break out of the way now.