Replace most SIMDE__*_C macros
The _C was really jsut meant as a short way to say "cast", but I always
felt like the name was a bit misleading because they are named
similarly to the macros in C like INT8_C, UINT8_C, INT16_C, etc., but
do not perform a similar function.  This name is a bit longer, but it's
more descriptive.