test: lots of work on cleaning up individual tests
This commit is mainly about getting rid of implicit conversions,
especially those C++ doesn't like, but there are other improvements
mixed in which are an effort to standardize the test layout; MMX
test improvements are particularly in need of such improvements as
they were the first ones added.  Later tests don't need to be touched
at all (notice there are no changes in AVX or AVX2 tests).

Another big part of this is converting hex constants to decimal.
I was using hex for a while because it's easier to find issues in
swizzle operations, but using hex for signed values can be problematic
so I'd rather just get rid of them.

Ther big remaining thing to be done is that there are a bunch of
places where we do stuff like

  simde_mm_set_epi16(1, 2, 3...)

Instead of

  simde_mm_set_epi16(INT16_C(1), INT16_C(2), INT16_C(3)...)

I'd like to eliminate all of those.