Use native types for arguments and return values
This is a *massive* patch, I'm sorry about that.  Unfortunately the
change requires a huge amount of code inside of SIMDe to also
change.  The public API doesn't change.

The idea is that instead of passing around our union, we should pass
a native vector if available (i.e., an __m128 on SSE, float32x4_t
for NEON, etc.), otherwise hopefully a GCC-style vector.  This provides
several major advantages:

 * Compilers are less likely generate extra moves and instead keep
   everything in vectors.  This isn't actually that big of a deal
   because compilers optimize out virtually all of this, but I have
   seen it miss a few.
 * Some compilers (MSVC 2017, I'm looking at you) don't allow you to
   pass around unions with increased alignment requirements, but seem
   to have hardcoded support for the vector types they know about.

Additionally, since we now have to convert to/from our union in the
function implementations, I was able to make the conversion functions
use memcpy and now we don't have type punning through unions!  Almost
every compiler was okay with the punning since we had the may_alias
attribute on all the members, but MSVC doesn't support may_alias.  In
the past I haven't actually had a problem with that since MSVC tends
to be extremely conservative about optimizations that break punned
code, but it was having problems with SIMDe.

That gets several compilers and platforms working and all tests
passing.  The only platform I've tried that now doesn't work is s390x,
and that's likely all endianness issues.  I'll have to look into it,
but IIRC Intel makes a lot of endianness assumptions in the API.

This is also a huge improvement for the native aliases.  Since we're
now using the native types instead of the SIMDe unions, you're able
to mix SIMDe and native implementations.  It also makes implementation
of aliases much simpler since we don't have to convert to/from the
SIMDe union in the alias macros, so bugs should be a bit easier to
spot.