Don't use -2147483648 in test vectors.
This can trigger warnings in some compilers since 2147483648 is too
large to fit in a 32-bit signed integer, and the sign isn't part of
the constant; the constant is the large positive, then it gets
negated.

I've always thought the sign was part of the constant.  Technically
I guess it would be more correct to use -INT32_C(X) instead of
INT32_C(X), but that seems really ugly to me…