avx2: don't use simde__m256i_to_private(*).m128i[*] in macros
Instead, use simde_mm256_extracti128_si256.  Without this, MSVC will
generate warning C4223 ("nonstandard extension used: non-lvalue array
converted to pointer").  No other compilers seem to have a problem
with this, and I believe it is allowed in C99 (which MSVC doesn't
support), so we could just suppress the warning.  However, I don't see
any disadvantage to this, and given the choice I'd rather preserve
the warning.

Since these are only in macros, they haven't been triggered so far.
The equivalent code in functions was switched over as part of
18dac2bec4460a9b007709e1560c95b4718d8402.  However, I'm about to merge
a patch which uses these from AVX-512F which does trigger the warning.