test/x86/skel: add b parameter to mm512_mask_xxx_* functions
This also modifies how we generate the masks to make sure that we don't
print a larger value than will fit in mask of the given size.  On
systems where uint_fast*_t != uint*_t, we want to make sure that our
generated masks will fit in the smaller (uint*_t).