common: remove SIMDE_CHECKED_*_CAST C++ implementations
Overloading is ambiguous when the types you're trying to differentiate
on are the same after accounting for typedefs.  This is likely going to
be a common case going forward.

What we really need here are opaque typedefs, which was proposed a few
years back but not adopted.  That leaves the option of either adding a
ton of code (nope) or using a third-party library (like the one
described by https://foonathan.net/2016/10/strong-typedefs/, which I
quite like and will probably be using in other code in the future).

I'm going to keep trying to find a way to do this safely in C++, but
meanwhile C++ users just don't get this feature.  In C++,
SIMDE_CHECKED_*_CAST will just be an uncoditional cast.  C will still
do the right thing.

Fixes #230