Remove double-underscores from include guards.
clang seems to think any macro names with two consecutive underscores
are reserved, at least in C++ mode.  I haven't bothered to check the
C++ standard, but even if this just works around a clang bug I'm
fine with it.

That said, since I'm also not adding a trailing underscore here this
moves the guards into the public SIMDe API.  This should make it
easy to detect if SIMDe has been included which allows code the
option of checking to see if SIMDe has been included and, if not,
include the native headers.  This allows SIMDe to optionally be used
by code without that code actually shipping with SIMDe.