sse, sse2: fall back on normal comparisons if fenv.h isn't present
Emscripten tot removed fenv.h, which makes sense since it isn't really
supported in WASM.  However, this was causing SIMDe to fall back on
HEDLEY_UNREACHABLE which caused CI to fail.  This doesn't quite do
what the functions are supposed to, but it's closer than completely
breaking.  And it gets CI working again.