features: try to undefine bool/vector/pixel on AltiVec
On ppc64el, altivec.h defines bool, pixel and vector which can collide with
c++ types.

As altivec.h explains, it's possible to undefine those for C++ compatibility
which let it define the variables that other files will use while not impacting
SIMDe.

Unfortunately, this seems to be GCC-specific.

Co-authored-by: Frédéric Bonnard <frediz@debian.org>

Fixes #269.