Rename macros with double underscores.
Clang seems to think this is disallowed by the C++ standard.  I don't
want to try to find it in the standard to verify that because it
doesn't matter; I want to compile cleanly with -Weverything even if
they're wrong, so I'm just converting private APIs to all end with
an underscore.