test: use SIMDE_MATH_NAN instead of SIMDE_MATH_NANF in codegen
glibc actually defines NAN to __builtin_nanf("") not __builtin_nan(""),
so the type is a float.  If you then assign it to a double while
-Wdouble-promotion is enabled (clang and GCC both support it), you end
up with a diagnostic.  SIMDE_MATH_NAN is defined to __builtin_nan("")
when supported; we fall back on NAN, but if need be we could add an
explicit cast to that.

I also regenerated the _mm_sqrt_{ps,pd} tests as a test for this.