sse: require MMX for native use of _m_pshufw
The problem here is that MSVC doesn't support MMX, or the 64-bit
vector functions in SSE, on AMD64.  To deal with this we check both
SIMDE_X86_SSE_NATIVE *and* SIMDE_X86_MMX_NATIVE (the latter of which
will be false on MSVC on AMD64) before calling the MMX-like SSE
functions.