travis: remove BUILD_TYPE, pass -O2 to ppc64 gcc
We have long since given up an BUILD_TYPE='Coverage'; Meson's support
for coverage builds actually works, so we use that instead.  We just
need to port the Travis builds over to Meson.

As for -O2, hopefully this fixes the PPC failures.  I'm working on a
reduced test case for GCC, and it looks like optimization fixes the
issue.