Add CONSTIFY macros.
This is a rather ugly hack to make it so we don't have to use integer
constant expressions when passing values to functions which the
compilers to be ICEs.  There should be no performance penalty **as
long as the value is a compile-time constant**.

We also use the same functionality in the tests to allow us to use the
normal test code instead of having to use an ICE.  This *will* incur
the run-time penalty, but in the tests we don't really care about that.
This makes the tests for these functions much easier to write.