test: fix test suite on emscripten
µnit expects test functions to have two parameters (one for parameters
and one for user data), but we don't use that functionality in SIMDe.
We just remove the parameters and cast the function pointers when we
assign them to the SIMDe suite.  I thought that was safe to do on all
platforms, but apparently it isn't allowed on emscripten.

Until I changed the test code recently every function had these
parameters, and we cast them both to void in every function.  Removing
the parameters and the casts cleaned things up a good amount, but I
didn't realize it broke emscripten.

To get around this, instead of the test functions taking no parameters
I defined a macro, SIMDE_MUNIT_TEST_ARGS, which expands to the params
µnit expects on some platforms (including emscripten), and `void`
elsewhere.  Right now we switch based on whether the compiler supports
the unused function attribute, but currently only check for GCC or
compilers based on clang (like emscripten), since we don't want to
trigger a ton of useless warnings.  It should be possible to expand
that list, or even default to on and add pragmas to suppress those
warnings, but I don't currently see much reason to do so.