travis: remove amalgamated build
We still push amalgamated builds out to Netlify, this is only for the
amalgamated branch, but that wasn't really doing any good anymore; it
was just a different way of getting a copy of the same data you can
download from Netlify, and was increasing the repo size for no good
reason.

It's also worth mentioning that there is also the simde-no-tests
repository, which contains only the implementations and is much better
suited for use as a submodule.