test: remove download-sde.py, add download-iig.sh
The download-sde.py script is unnecessary and no longer works.  Intel
changed their site like 2 days after I wrote it.  The good news is
that this is *much* easier with the new site…  the necessary commands
are embedded in .github/workflows/ci.yml if you need them.

The script no download the IIG is probably overkill, but the IIG data
is handy for various scripts so a convenient way to grab a copy
without running native-aliases.sh is nice.