neon: Add implementation of reinterpret
this is implemented via memcpy, which, honestly, is a pretty god-awful
way of doing this. However, the way I wrote the file means that if
there's a way to do this better, all that I have to do is swap out
three lines and it generates the code properly.

removing whitespace

Completely rewrite the generation of the reinterpret functions