docker: move apt-get update to different command
Docker caches the results of each command pretty aggressively.  So if
we do an apt-get update at the end of a command it will likely stick
around for a long time, even if we're trying to install at a much later
date.  This just moves the update into the same command as the install
so we install from an up to date repo.