x86: indentation fixes
With this I believe SIMDe should be completely migrated to the "new"
style where proprocessor directives are indented just like regular
code, and increase the indentation of regular code.

Sorry it's such a massive and disruptive change; I *really* should have
done this much earlier.  I thought it would be better to just clean up
each function as we modified it, but that didn't really happen and more
functions were added using the old style.  Hopefully that ends now that
there is a CI check to test for at least some of those.