Fix compilation failures when targeting 32-bit x86 with >= SSE2
There are two problems here.  The primary issue is that several
functions are not available on x86, and it isn't obvious from the
Intel Intrinsics Guide which ones.  Adding -march=native to our x86
Travis build should help us pick up on those, but unfortunately even
though it will build with this patch some tests fail.

The second issue is GCC bug #53784.