sse: add __builtin_nontemporal_store version of simde_mm_stream_ps
The builtin is only available on clang, but it does exactly what we
want.  The stream functions are supposed to differ from the store
functions in that they are non-temporal, but so far we've really just
been treating them as aliases.

On many non-x86 CPUs I think they basically will be aliases still;
AFAIK Arm doesn't have non-temporal vector stores, for example.  Still
at least this gives the compiler a good optimization hint in case there
is an architecture that does support it (other than x86, where we just
use _mm_stream_ps anyways).