sse4.1: improve AArch64 _mm_dp_{ps,pd} implementations
The basic idea is to take advantage of vaddvq_f{32,64} instead of
pulling values out of the vectors, performing an operation, then
inserting the results into a new vector.  To do this we just have
to generate a couple of masks based on the imm8 parameter and use
them to mask out the unwanted values.

I also regenerated the test cases so we hit each code path.