meson, cmake: add -DSIMDE_FAST_MATH to C/C++ flags on ICC
ICC enables unsafe optimizations by default; to get precise results
you have to pass -fp-model=precise (or a few other possible values).
However, unlike GCC and Clang, there doesn't seem to be any way to
determine whether the compiler is using fast or precise math.

This patch is only about the tests; we still assume precise math in
the code, but since most people building with ICC will be using fast
math that's what we'll test.  The other possibility would have been
to add -fp-model=precise to the flags, but that seemed a slightly
worse "solution" to me.