gfni: lower requirements for some functions
On Tremont, Intel supports 128-bit GFNI functions even though the CPU
does not support AVX-512.

According to the latest SDM (the IIG hasn't been updated), only GFNI is
required for the legacy encoded instructions.  AVX is required for
VEX encoded instructions, and AVX-512 is required for EVEX.

Basically, that maps to GFNI for 128-bit functions, AVX for 256 bit
functions, and AVX-512F for 512-bit functions, with the added caveat
that the mask/maskz functions on 128/256-bit vectors require AVX-512VL.

This is based on a question I asked on Intel's forums at:
https://community.intel.com/t5/Intel-ISA-Extensions/Requirements-for-GFNI-without-AVX-AVX-512/m-p/1228325/thread-id/6852#M6853
The relevant bits of the SDM are volume 2A, 3-442, 3-445, and 3-447.