clmul: don't use __builtin_shufflevector on XLC
XL C/C++ supports __builtin_shufflevector, but only when using AltiVec
types, not on GCC-style vector extensions.  We could add alternate
implementations which use the same code but (for example) altivec_f32
instead of f32, but that introduces a lot of bloat into the code for
a compiler very few use *and* rewards IBM for a stupid decision which
breaks compatibility with *lots* of code.