sse4.1: regenerate _mm_dp_ps test
With the previous test some values were large enough to cause
sufficient enough differences that the test would fail on some
platforms (specifically, i686 with optimizations enabled).  The results
weren't really incorrect; I believe the difference came down to the
order of the calculation for the sum; Intel uses ((0 + 1) + (2 + 3)),
which is what _mm_hadd_ps does, but without SSE the compiler may reduce
it in another order.

Fixes #648.