docker: replace clang-8 cross file with one for clang-11
clang-8 is no longer included in Debian testing.  The container still
builds fine since we don't hard-code clang versions in it, and instead
just install all the versions we can find.  However, the cross file
are still hard-coded.

The long-term fix for this is to replace the handwritten cross files
with a script to automatically generate cross files based on the
installed compilers.  I'm planning to do this, but it requires a
fair amount of work since there are a lot of complications.

Fixes #666