xop: initial implementation
XOP was an AMD ISA extension which Intel never adopted.  It is present
on heavy-equipment CPUs (bulldozer, piledriver, etc.), but no longer
supported on Zen.  It never really attained widespread use, but this
could still be useful.

Huge thanks to Jan Ringoš (@tringi) for testing this during
development; I don't have access to any hardware that supports XOP
so Jan's involvement was absolutely critical.

This is just the portable version; we still need to go through and add
accelerated implementations.

Fixes #678