drone: adjust branch triggers
In the old configuration Drone.io was only running on ci/drone.  What
we want is for it to run on every branch except for master or ci/*
(other than ci/drone and ci/drone/*, which we *do* want to run).

This just explicitly lists the other CI providers we want to ignore
so we run everything in ci/* other than those specific branches.  It's
a bit fragile, but it should mostly do the trick.