avx: try to detect prior inclusion of AVX header and handle it
This handles the case where AVX support is not enabled, but the AVX
header has been included.  In that case there will be typedefs for the
__m256/__m256i/__m256d types which we can't use.

In order to detect such a situation we look at a few different macros:
GCC defines _AVXINTRIN_H_INCLUDED, clang defines __AVXINTRIN_H.  We
also check for _CMP_EQ_OQ, but I don't want to rely on that since it
may be implemented on some compilers as an enum value instead of a
macro.

If any of those macros are defined, we assume that AVX support has
already been included and use macros for the __m256 types instead of
typedefs.