gh-actions: temporarily disable emscripten build
There is a bit of turmoil right now as projects move to the final WASM
SIMD opcodes; LLVM now uses them (at least in tot), but d8 and wasmtime
don't, so we have no way to execute the tests.  The final opcodes are
supported in Chrome canary, but I don't think there has been a v8
release yet.

I'll re-enable this as soon as either d8 or wasmtime can handle it.