neon: begin refactoring to use different types on all targets
Currently, we typedef native types to all of the NEON types.  For
WASM and x86, this means multiple NEON types are typedefd to the same
type (e.g., simde_int8x16_t and simde_int16x8_t are both typedefd to
__m128i on x86, and v128_t on WASM).  This causes problems in some
software, for example C++ functions overloaded based on the NEON type.

In order to fix this, we have to refactor the code so the other
versions use the *_private structs after we have converted them.  Once
that's done we can switch the public types to whaever we want, such as
vector extension types on x86 for most compilers, or the *_private
types for MSVC.