avx, avx512/cmp: properly handle NaN in _mm{,256,512}_cmp_{ps,pd,ss,sd}
This pretty much amounts to a rewrite of these functions and the
associated tests, but the result should be much better.

There are still a few variants I need to rewrite the tests for, so I've
left the test case generation code in place, but since the core
functions are properly tested now I'm a lot less concerned.