Many work-arounds for GCC with MSA, and support in the docker image.
With this, we can actually compile on GCC when targeting MSA.  Several
tests still fail, but it appears to more or less work.

Unfortunately QEMU 5.2.0 doesn't work due to MSA support not being
included in the emulated CPU (though the feature flag is set), so the
version from Debian testing is out, and 6.0.0 seems to have some
buggy implementations resulting in incorrect values, so we have to
build from git.  This is a lot of data to download and compile, so it's
off by default; set QEMU_GIT=y when calling simde-dev.sh to enable it.