gh-actions: remove GCC 4.7 build
The 4.7 build has been a major problem lately; performance is really
bad.  It's just not feasible to continue testing 4.7 on CI; I'll try
to keep it working as best I can, but it will have to be with
periodic manual testing on a machine with more resources (or less
parallelism in the build).