simd128: move tests from wasm/ to wasm/simd128/
This is just to make room for relaxed-simd and (eventually) flexible
vectors.

This patch also fixes µnit/CMake support in simd128 so the tests will
actually run when using cmake.  Apparently this was not the case
before.  Oops.  To be clear, though, they were running with meson
which is what we usually use to build.