simd128: work around bad diagnostic from clang < 7
The old version just used vector extensions to set the index, but
the return value was the internal type for vectors with that element
type, so some versions would trigger a -Wvector-conversion diagnostic
when assigning the return value to a __m128i.

This was fixed (unintentionally) in SVN 334057 (git f3914b74) when
they added versions which use intrinsics so they could limit the
inputs to constants.