travis:  use -march=native and GCC on s390x
There are still some tests which fail with clang, but on GCC everything
works.  The hardware seems to be limited, at least sometimes, to z13 so
z14-specific instructions were obviously a problem as well.

With this we can move the tests into the script portion so when they
fail the job fails, which should help us keep the ports working.