Default to -DSIMDE_CONSTRAINED_COMPILATION when building tests
The tests tend to include a lot of repeated calls to the same function
so when they are testing a large function it really hurts.  With this
we just don't force the compiler to inline these giant functions,
which slows down the implementation a bit at runtime but doesn't alter
the program so the tests are still valid.  We still want to default
to aggressive inlining when *not* running the tests, though.