Allow non-immediate values in _mm_slli_epi api
_mm_slli_epi32 and their ilk allow non-constant values, for instance
like `int shift=2; _mm_slli_epi32(a, shift);`. The SIMDE translation
of these apis to `vshlq_n_s32` and so on required shift to be immediate
mode (functionally constexpr) meaning valid SSE2 code would not compile.

This diff changes the API to `vshlq_s32` with the associated broadcast
of the shift register, adds tests, removes the constant constraint on the
API. All tests pass on my macos 12.4 / xcode 14 M1 macbook.