circle-ci: modernize build, especially for i686 & loongson
debian:*slim doesn't have git or ssh which leads to a warning:
"Either git or ssh (required by git to clone through SSH) is not
installed in the image. Falling back to CircleCI's native git client but
the behavior may be different from official git. If this is an issue,
please use an image that has official git and ssh installed."