Don't use _Float16 on architectures that do not support it
_Float16 is only supported on a few architectures. Let's assume it's not
supported unless we know otherwise. Then we no longer need to check for
and exclude MIPS or POWER, and don't need to add more exclusions for
missing architectures that need to be excluded (notably s390x). Notably,
both GCC and Clang documentation lists the architectures where this type
*is* supported, and that list is small in both cases.

_Float16 is also not supported on x86 unless SSE2 is available, so let's
tighten the check for that architecture as well.

References:

https://bugs.webkit.org/show_bug.cgi?id=274086
https://gcc.gnu.org/onlinedocs/gcc/Half-Precision.html
https://clang.llvm.org/docs/LanguageExtensions.html#half-precision-floating-point