x86 sse2: small fixes for loongarch
At present, the instructions like slli/srli on loongarch only
accept immediate constant values for shit count, so in order to
avoid compiler errors in the build process, we need to use
sll/srl/sra version instead. In the feature, if the compiler
supports the formers, we can restore them.